+++ core/org.eclipse.cdt.core/parser/org/eclipse/cdt/internal/core/pdom/PDOMFileSet.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2008 Wind River Systems, Inc. and others.
+ * Copyright (c) 2008, 2009 Wind River Systems, Inc. and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -21,7 +21,7 @@ import org.eclipse.core.runtime.CoreException;
 
 public class PDOMFileSet implements IIndexFragmentFileSet {
 
-	private HashSet<Integer> fFileIDs= new HashSet<Integer>();
+	private HashSet<Long> fFileIDs= new HashSet<Long>();
 	
 	public void add(IIndexFragmentFile fragFile) {
 		PDOMFile pdomFile= (PDOMFile) fragFile;
