+++ core/org.eclipse.cdt.core/parser/org/eclipse/cdt/internal/core/pdom/db/IBTreeComparator.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2005 QNX Software Systems and others.
+ * Copyright (c) 2005, 2009 QNX Software Systems and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -21,6 +21,6 @@ public interface IBTreeComparator {
 	/**
 	 * Compare two records. Used for insert.
 	 */
-	public abstract int compare(int record1, int record2) throws CoreException;
+	public abstract int compare(long record1, long record2) throws CoreException;
 	
 }
