+++ core/org.eclipse.cdt.core/parser/org/eclipse/cdt/internal/core/pdom/db/IString.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2006, 2007 QNX Software Systems and others.
+ * Copyright (c) 2006, 2009 QNX Software Systems and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -25,7 +25,7 @@ public interface IString {
 	/**
 	 * Get the offset of this IString record in the PDOM
 	 */
-	public int getRecord();
+	public long getRecord();
 	
 	// strcmp equivalents
 	/**
