+++ core/org.eclipse.cdt.core/parser/org/eclipse/cdt/internal/core/pdom/dom/cpp/PDOMCPPFunctionType.java
@@ -67,7 +67,7 @@ public class PDOMCPPFunctionType extends PDOMCFunctionType implements ICPPFuncti
 	IPointerType thisType;  // Cached value
 	int cvq= -1; // Cached value
 	
-	protected PDOMCPPFunctionType(PDOMLinkage linkage, int offset) {
+	protected PDOMCPPFunctionType(PDOMLinkage linkage, long offset) {
 		super(linkage, offset);
 	}
